Simplify delete button event listeners
The delete buttons weren't working properly. After deleting one
book, the other delete buttons would not work. Previously event
listeners were assigned with a complicated for loop which seemed
to be the problem. The for loop was used because I thought I had
to select the created element after it was appended to its parent.
Instead, event listeners can be added before.

[Ticket: X]